refactor

+++ modules/pi_estimator.py
@@ -1,2 +1,5 @@
-def python_estimate_pi(samples: int) -> float:
+from numpy import float64, uint64
+
+
+def python_estimate_pi(samples: uint64) -> float64:
     return 3
